refactor(7141): isolate MCP session configuration
Move child environment policy and timeout validation out of the persistent transport session, keeping both modules below the repository size limit and adding safe millisecond overflow rejection.

Constraint: Public imports continue through mcp-session.ts so extension callers do not change.

Rejected: Compress configuration logic inside the session module | it would keep transport and spawn policy coupled at the file limit.

Confidence: high

Scope-risk: narrow

Directive: Keep credential filtering and deadline parsing together in the MCP child configuration module.

Tested: all 61 KAMN Pi extension tests; mcp-session.ts 147 lines; mcp-session-config.ts 68 lines; diff check.

Not-tested: Fresh live integration is the next phase.

Related: #7141
